Split off actions pulling in ogc-schemas into a separate file
